fix: not remove xdebug when coverage is enabled

+++ tests/bin/php/setup-wp-tests.sh
@@ -2,7 +2,9 @@
 
 set -ex
 if [[ -f ~/.phpenv/versions/$(phpenv version-name)/etc/conf.d/xdebug.ini ]]; then
-	phpenv config-rm xdebug.ini
+	if [[ -z "$COVERAGE_REPORT" ]] ; then
+		phpenv config-rm xdebug.ini
+	fi
 else
 	echo "xdebug.ini does not exist"
 fi
